requirements.txt
django==3.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-cors-headers==3.12.0
python-dotenv==0.19.2
django-extensions==3.1.5
django-filter==21.1
django-celery-beat==2.3.0
django-celery-results==2.4.0
drf_writable_nested
requests
django-celery-beat==2.3.0
django-celery-results==2.4.0
Redis
drf-yasg